microlearn/requirements.txt
django>=5.0,<7.0
djangorestframework>=3.15
django-cors-headers>=4.0
requests>=2.30.0
google-generativeai>=0.8.0
